增加了文件分配器。

+++ vtos/port/os_cpu_def.h
@@ -10,8 +10,8 @@ typedef unsigned short uint16;
 typedef signed short int16;
 typedef unsigned int uint32;
 typedef signed int int32;
-typedef unsigned long uint64;
-typedef signed long int64;
+typedef unsigned long long uint64;
+typedef signed long long int64;
 typedef unsigned int os_stk; //堆栈类型，堆栈存取宽度
 typedef unsigned int os_cpu_sr; //用于保存CPU状态寄存器类型
 typedef unsigned int os_size_t; //这个类型要保证能放下一个地址
